feat(procurement): full Edit-PO editor inline in the workspace (shared component)
Extracted the big inline Edit-PO modal out of the 1361-line orders page into a reusable
<EditOrderModal> (invoice upload + AI extract, invoice number/dates, deposit %, editable
line items, Save / Upload-Invoice-&-Send). Single source of truth, used in both places.

- components/inventory/EditOrderModal.tsx — owns all edit state/handlers/JSX; props
  order/confirmMode/onClose/onSaved; faithful move, no logic/API changes.
- orders/page.tsx — inline modal + ~20 state vars + handlers removed (−859 lines); now
  just a setEditOrder trigger + <EditOrderModal>; behaves identically.
- supplier-chats workspace — PO panel gets an "Edit" button → opens the full editor
  inline (fetch orders/[id] → adapt → modal), refreshing on save. Status actions kept.
- orders/[id] GET now includes invoices (orderBy createdAt desc) so an existing invoice
  pre-fills when editing from the workspace (closes the parity gap).

Extraction verified at typecheck + lint (0 errors). Touches the critical orders page —
worth a click-through on the preview deploy.
